Scheduling_and_transport/backend/requirements.txt

Django==5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
whitenoise==6.6.0
gunicorn==21.2.0
reportlab==4.0.4
Pillow==10.0.1
requests==2.31.0
psycopg2-binary==2.9.7
dj-database-url
